backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Redis
redis==5.2.1

# OpenAI
openai==1.57.4

# Export
reportlab==4.2.5
python-docx==1.1.2

# MinIO
minio==7.2.13

# HTTP client
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1

# Dev tools
python-dotenv==1.0.1
